✅ MD4XTG code: done
Task: 202604231744-MD4XTG
Primary: code
Agent: CODER
Author: CODER
Status: DONE
Comment: verified: split the README migration logic out of migrate-doc.ts, kept the command orchestration stable, and confirmed the hotspot gate, package typecheck, and patch hygiene stay green for the touched task-doc migration path.
